Implemented sorting. Started work on code cleanup and intergration with UI

+++ SBD Proj 1/uiTable.cpp
@@ -174,7 +174,7 @@ unsigned int uiTable::getContentHeight() const {
 	return this->position.H - 1;
 }
 unsigned int uiTable::getContentSize() const {
-	return this->contentField.size();
+	return (unsigned int)this->contentField.size();
 }
 unsigned int uiTable::getContentOffset() const {
 	return this->contentOffset;
